fix: right type for the port property
[#184098479](https://www.pivotaltracker.com/story/show/184098479)

+++ integration-tests/postgre_sql_test.go
@@ -81,7 +81,7 @@ var _ = Describe("PostgreSQL", Label("PostgreSQL"), func() {
 				"password": "bind.test.password",
 				"uri":      "bind.test.uri",
 				"use_tls":  true,
-				"port":     5443,
+				"port":     float64(5443),
 			}))
 		})
 	})
